requirements.txt
asgiref==3.3.1
Django==3.0.6
pytz==2020.1
six==1.14.0
psycopg2-binary==2.8.5
django-filter==2.4.0
django-widget-tweaks==1.4.5

